requirements.txt
numpy==1.18.1
tesserocr==2.5.1
opencv_python==4.5.1.48
Pillow==8.1.2
pydub==0.25.1
moviepy==1.0.3
